+++ views/dashboard.handlebars
@@ -18,7 +18,7 @@
 
             </div>
             <div class="medium-6">
-                <form class="grid-x" id="sort" action="/adSortOrder" method="post">
+                <form class="grid-x" id="sort" action="/dashboard/adSortOrder" method="post">
                     <div class="cell medium-8">
                         <div class="sort-by-text">
                             <label for="sort-order">Sort Ads by: </label>
